:sparkles: add: When deleting a specific HUB, HUB that has already been soft deleted cannot be deleted.

+++ com.rush.logistic.client.hub/src/main/java/com/rush/logistic/client/hub/service/HubService.java
@@ -100,6 +100,13 @@ public class HubService {
                     .orElseThrow(() ->
                             new IllegalArgumentException(HubMessage.HUB_NOT_FOUND.getMessage())
                     );
+
+            // soft delete된 허브입니다.
+            if (hub.isDelete()) {
+                return BaseResponseDto
+                        .<HubIdResponseDto>from(HttpStatus.GONE.value(), HttpStatus.GONE, HubMessage.HUB_ALREADY_DELETED.getMessage(), null);
+            }
+
             // 허브 삭제정보 업데이트
             hub.delete();
